# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Healthcare Data Mapping Tool API",
    "description": "\nv1 REST API for the Healthcare Data Mapping Tool.\n\n## Auth (v1 simulated RBAC)\nPass `X-User-Id` and `X-User-Role` headers with every request.\nRoles: `viewer`, `analyst`, `approver`, `admin`.\n\n## Phase scope\n- Schema import + catalog (Phase 3)\n- Mapping suggestions (Phase 4)\n- Mapping validation with configurable policy (Phase 4)\n- Mapping artifact export (Phase 6 script placeholder)\n\n## Out of scope for v1\n- Transformation execution (Phase 6)\n- Patient-level data ingestion (forbidden in v1)\n",
    "version": "1.0.0"
  },
  "paths": {
    "/projects": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "create_project_projects_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_projects_get",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project Status",
        "operationId": "update_project_status_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/suggestions": {
      "post": {
        "tags": [
          "suggestions"
        ],
        "summary": "Get Suggestions",
        "description": "Generate TOP-K mapping suggestions for every target field.\n\nUses pinned source & target schemas — no raw schema parsing here (RL-05).\nNo patient data is accessed (RL-02).",
        "operationId": "get_suggestions_projects__project_id__suggestions_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/validate": {
      "post": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Mapping",
        "description": "Validate a mapping version against pinned schemas with optional policy overrides.",
        "operationId": "validate_mapping_projects__project_id__validate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sensitivity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensitivity",
            "default": "healthcare_highly_sensitive"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateProjectRequest"
      },
...
      "FieldSuggestionsDTO": {
        "properties": {
          "target_field_id": {
            "type": "string",
            "title": "Target Field Id"
          },
          "target_field_path": {
            "type": "string",
            "title": "Target Field Path"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/SuggestionCandidateDTO"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "type": "object",
        "required": [
          "target_field_id",
          "target_field_path",
          "candidates"
        ],
        "title": "FieldSuggestionsDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "sensitivity": {
            "type": "string",
            "title": "Sensitivity"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "owner",
          "sensitivity",
          "created_at"
        ],
        "title": "ProjectResponse"
      },
...
      "SeverityOverride": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "description": "error | warn | info"
          }
        },
        "type": "object",
        "required": [
          "rule_id",
          "severity"
        ],
        "title": "SeverityOverride"
      },
      "SuggestionCandidateDTO": {
        "properties": {
          "source_field_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Field Ids"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source_field_ids",
          "confidence",
          "reasons",
          "warnings"
        ],
        "title": "SuggestionCandidateDTO"
      },
      "SuggestionsRequest": {
        "properties": {
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "min_confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Confidence",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "SuggestionsRequest"
      },
      "SuggestionsResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "source_schema_id": {
            "type": "string",
            "title": "Source Schema Id"
          },
          "target_schema_id": {
            "type": "string",
            "title": "Target Schema Id"
          },
          "field_suggestions": {
            "items": {
              "$ref": "#/components/schemas/FieldSuggestionsDTO"
            },
            "type": "array",
            "title": "Field Suggestions"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "source_schema_id",
          "target_schema_id",
          "field_suggestions"
        ],
        "title": "SuggestionsResponse"
      },
...
      "UpdateProjectStatusRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Target status. Allowed: review, approved, published, archived"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "UpdateProjectStatusRequest"
      },
      "ValidateRequest": {
        "properties": {
          "mapping_version_id": {
            "type": "string",
            "title": "Mapping Version Id"
          },
          "severity_overrides": {
            "items": {
              "$ref": "#/components/schemas/SeverityOverride"
            },
            "type": "array",
            "title": "Severity Overrides"
          }
        },
        "type": "object",
        "required": [
          "mapping_version_id"
        ],
        "title": "ValidateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationIssueDTO": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "affected_field_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Field Ids"
          },
          "affected_field_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Field Paths"
          },
          "remediation_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation Hint"
          }
        },
        "type": "object",
        "required": [
          "rule_id",
          "severity",
          "message",
          "affected_field_ids",
          "affected_field_paths",
          "remediation_hint"
        ],
        "title": "ValidationIssueDTO"
      },
      "ValidationResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "mapping_version_id": {
            "type": "string",
            "title": "Mapping Version Id"
          },
          "has_errors": {
            "type": "boolean",
            "title": "Has Errors"
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count"
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueDTO"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "mapping_version_id",
          "has_errors",
          "error_count",
          "warning_count",
          "issues"
        ],
        "title": "ValidationResponse"
      }
    }
  }
}